+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add conversation history size setting
+- Add enable reply moderation support
 
 ### Changed
 - Change default model to gpt-5.2